requirements.txt
fastapi[all]           # FastAPI + Starlette + Pydantic
uvicorn[standard]      # Dev server (with live reload)
sqlalchemy[asyncio]    # Async database ORM
asyncpg                # Async PostgreSQL driver
alembic                # DB migrations
python-dotenv          # Load .env file for configs
python-jose[cryptography]
passlib[bcrypt]
python-multipart       # needed by OAuth2PasswordRequestForm
email-validator        # used by Pydantic EmailStr
pydantic-settings      # Step 1/2 settings
